docs: record that a skill band below the floor destroys dispersion
Measured in-game 2026-08-07, 1000 pawns per profile, active profile only with
no overrides in play.

Wildcard's per-pawn skill dispersion (sd 1.10) is NARROWER than Faithful's
(1.23). The preset whose entire purpose is maximum variation delivers less
pawn-to-pawn skill variation than the vanilla-like baseline. Its median per-skill
level is 0.0, and 506 of 1000 pawns averaged under 0.6 across all twelve skills.

Mechanism: skillShiftMin = -8.7 is applied on top of vanilla's own levels, which
average ~3.4 (Faithful's column is effectively a vanilla readout, since it barely
shifts anything). Most rolls land below zero and Shift's Clamp(0, 20) stacks them
onto the same floor. There is no spread below a wall you are already pressed
against -- so the lower half of the distribution becomes a point mass at 0. That
is the same "a spread becomes a spike" failure already documented for the
rejected downside-clamp proposal, arriving through the band instead.

Why this needed writing down rather than just fixing: NOTHING THIS PROJECT
COMPUTES OFFLINE CAN SEE IT. CalculateCompositeScore and envelope_check.py both
read Lerp(skillShiftMin, skillShiftMax, q) and neither models the per-pawn clamp,
because neither generates a pawn. Wildcard passes Rule 1 comfortably at -20.8%
while its population is censored. The dispersion table is equally blind: it
reports 2.08 lv for Wildcard, 4x Faithful, which is true of the intended noise
and false of what ships. A preset can be inside the envelope, inside the
dispersion table, and broken in the only place that matters.

Records four rules on the band, at the paragraph that previously said only that
pinning was "a tuning outcome": keep skillShiftMin above roughly -4; widening a
band downward past the floor REDUCES dispersion; never judge a skill band from
envelope_check.py alone -- read the MEDIAN from the in-game dump; and this is a
band property, not a noise one, so narrowing skillNoise would not fix it.

Same warning on the constant itself in VarianceProfile.cs, since that is where
someone reaching for "more variance" will actually be typing.

+++ Source/TraitVarianceApplier.cs
@@ -233,7 +233,7 @@ namespace PawnVarianceMod
 
         private static float JitterSample()
         {
-            return ((float)Rand.Value - 0.5f) * Constants.SmallRandomJitter;
+            return (Rand.Value - 0.5f) * Constants.SmallRandomJitter;
         }
     }
 }
